added ability to used stored credentials
this functionality loads secrets from secrets.yml (example provided). If no secrets file is found it behaves as before. If one is found, accounts are loaded. When account is selected, the api client name and key get loaded. If you wish enter a new account and credential set, select the last item in account "enter manually" and you can then enter as you normally would and save if you wish to do so. Then the credential will be loaded on the next load of accounts for you to use again.

This is a huge time saver for SE's, CX and partners who use this script regularlarly! Sorry I didn't include this from the get go!