added servers and enriched node data
now there is a servers sheet with a dump of everything from that API. I am now merging servers data with nodes data to enrich the nodes sheet with relevant info from the servers data. Provided there are machine agents on nodes with app agents running, we will get a picture of if the node is a container, cpu core counts, memory and more. This can be used to figure out IBL licensing and mciroservices licensing as well.